refactor: HelpTab.tsx を分割（582行 → 複数ファイル）
Issue #253 の要件に従い、HelpTab.tsx を以下のコンポーネントに分割:
- HelpGettingStarted.tsx: Getting Started セクション
- HelpFAQ.tsx: FAQ セクション
- HelpTroubleshooting.tsx: Troubleshooting セクション
- HelpDataPrivacy.tsx: Data & Privacy セクション
- HelpSettingsBackup.tsx: Settings Backup セクション（エクスポート・インポート機能）

HelpTab.tsx はレイアウト管理のみを担当（152行に縮小）
言語切り替え機能は HelpTab に残している

変更前: 582行
変更後: 152行（HelpTab） + 5つの個別コンポーネント
